Move check_specs_by_output_name into InOutMapper
